added money update after buying the share, added share to graphql query

+++ src/model/db/Players.kt
@@ -4,7 +4,7 @@ import org.jetbrains.exposed.sql.Table
 
 object Players: Table(){
     val id = integer("id").uniqueIndex().primaryKey().autoIncrement()
-    val money = integer("money")
+    val money = float("money")
     val userId = integer("user_id") references Users.id
     val startedAt = date("started_at")
     val removedAt = date("removed_at").nullable()
