+++ src/model/db/Transactions.kt
@@ -8,7 +8,7 @@ object Transactions: Table() {
     val id = integer("id").uniqueIndex().primaryKey().autoIncrement()
     val playerId = integer("player_id") references Players.id
     val companyId = integer("company_id") references Companies.id
-    val pricePerShare = integer("price_per_share")
+    val pricePerShare = float("price_per_share")
     val quantity = integer("quantity")
     val createdAt = date("created_at")
     val type = enumeration("type", TransactionType::class)
